# backend/api/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "invoice/v1/invoice.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/invoices": {
      "get": {
        "operationId": "InvoiceService_ListInvoices",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListInvoicesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "InvoiceService"
        ]
      },
      "post": {
        "operationId": "InvoiceService_CreateInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateInvoiceRequest"
            }
          }
        ],
        "tags": [
          "InvoiceService"
        ]
      }
    },
    "/v1/invoices/{id}": {
      "get": {
        "operationId": "InvoiceService_GetInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "InvoiceService"
        ]
      },
      "delete": {
        "operationId": "InvoiceService_DeleteInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "InvoiceService"
        ]
      }
    },
    "/v1/invoices/{invoice.id}": {
      "put": {
        "operationId": "InvoiceService_UpdateInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "invoice.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InvoiceServiceUpdateInvoiceBody"
            }
          }
        ],
        "tags": [
          "InvoiceService"
        ]
      }
    },
...
  },
  "definitions": {
    "InvoiceServiceUpdateInvoiceBody": {
      "type": "object",
      "properties": {
        "invoice": {
          "type": "object",
          "properties": {
            "orderId": {
              "type": "string"
            },
            "totalAmount": {
              "type": "string"
            },
            "paidAmount": {
              "type": "string"
            },
            "dueDate": {
              "type": "string",
              "format": "date-time"
            },
            "paymentDate": {
              "type": "string",
              "format": "date-time"
            },
            "paid": {
              "type": "boolean"
            },
            "items": {
              "type": "array",
              "items": {
                "type": "object",
                "$ref": "#/definitions/v1InvoiceItem"
              }
            }
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateInvoiceRequest": {
      "type": "object",
      "properties": {
        "invoice": {
          "$ref": "#/definitions/v1Invoice"
        }
      }
    },
    "v1CreateInvoiceResponse": {
      "type": "object",
      "properties": {
        "invoice": {
          "$ref": "#/definitions/v1Invoice"
        }
      }
    },
...
    "v1DeleteInvoiceResponse": {
      "type": "object"
    },
...
    "v1GetInvoiceResponse": {
      "type": "object",
      "properties": {
        "invoice": {
          "$ref": "#/definitions/v1Invoice"
        }
      }
    },
...
    "v1Invoice": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "orderId": {
          "type": "string"
        },
        "totalAmount": {
          "type": "string"
        },
        "paidAmount": {
          "type": "string"
        },
        "dueDate": {
          "type": "string",
          "format": "date-time"
        },
        "paymentDate": {
          "type": "string",
          "format": "date-time"
        },
        "paid": {
          "type": "boolean"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1InvoiceItem"
          }
        }
      }
    },
    "v1InvoiceItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "invoiceId": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "unitPrice": {
          "type": "string"
        },
        "totalCost": {
          "type": "string"
        }
      }
    },
    "v1ListInvoicesResponse": {
      "type": "object",
      "properties": {
        "invoices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Invoice"
          }
        }
      }
    },
...
    "v1UpdateInvoiceResponse": {
      "type": "object",
      "properties": {
        "invoice": {
          "$ref": "#/definitions/v1Invoice"
        }
      }
    },
...
  }
}